refactor: lotto 패키지 사용
