+++ src/main/kotlin/lotto/Application.kt
@@ -1,0 +1,8 @@
+package lotto
+
+import lotto.controller.LottoController
+
+fun main() {
+    val lottoController = LottoController()
+    lottoController.run()
+}
