+++ src/main/kotlin/lotto/model/GameState.kt
@@ -1,0 +1,7 @@
+package lotto.model
+
+sealed class GameState {
+    object Play : GameState()
+
+    object End : GameState()
+}
