+++ src/main/kotlin/lotto/model/Lottos.kt
@@ -1,0 +1,7 @@
+package lotto.model
+
+data class Lottos(val publishedLottos: List<Lotto>) {
+    override fun toString(): String {
+        return publishedLottos.joinToString("\n")
+    }
+}
