+++ src/main/kotlin/lotto/util/LottoConstants.kt
@@ -1,0 +1,9 @@
+package lotto.util
+
+object LottoConstants {
+    const val UNIT_PRICE = 1_000
+    const val START_RANGE = 1
+    const val END_RANGE = 45
+    const val SIZE = 6
+    const val MINIMUM_PURCHASE_SIZE_OF_MANUAL_LOTTO = 0
+}
