--- src/main/kotlin/model/GameState.kt
@@ -1,7 +1,0 @@
-package model
-
-sealed class GameState {
-    object Play : GameState()
-
-    object End : GameState()
-}
