--- src/main/kotlin/model/Lottos.kt
@@ -1,7 +1,0 @@
-package model
-
-data class Lottos(val publishedLottos: List<Lotto>) {
-    override fun toString(): String {
-        return publishedLottos.joinToString("\n")
-    }
-}
