--- src/main/kotlin/util/MoneyConstants.kt
@@ -1,10 +1,0 @@
-package util
-
-object MoneyConstants {
-    const val FIRST_RANK_WINNING_MONEY = 2_000_000_000
-    const val SECOND_RANK_WINNING_MONEY = 30_000_000
-    const val THIRD_RANK_WINNING_MONEY = 1_500_000
-    const val FOURTH_RANK_WINNING_MONEY = 50_000
-    const val FIFTH_RANK_WINNING_MONEY = 5_000
-    const val MISS_WINNING_MONEY = 0
-}
